Rename regression test to snapshot test
