+++ .travis.yml
@@ -13,7 +13,7 @@ script:
   - dpkg --info ../*.deb
   - dpkg --contents ../*.deb
   - sudo build_scripts/travis-install-and-setup-deb.sh
-  - py.test-3 ./parser_tests -vv
+  - py.test-3 ./snapshot_tests -vv
   - wget --output-document=/dev/null --input-file=build_scripts/test-urls.txt
   - wget --output-document=/dev/null --input-file=build_scripts/maybe-urls.txt || true
 after_success:
